+++ app/src/test/kotlin/com/stevenfrew/beatprompter/mock/MockPreferences.kt
@@ -94,7 +94,8 @@ class MockPreferences(
 	override val trimTrailingPunctuation: Boolean = false,
 	override val useUnicodeEllipsis: Boolean = false,
 	override val beatCounterHeight: Int = 4,
-	override val includeVariationsInFilterList: Boolean = false
+	override val includeVariationsInFilterList: Boolean = false,
+	override val sendInitialMidiMessagesAtTitleScreen: Boolean = true
 ) : Preferences {
 	override fun getStringPreference(key: String, default: String): String {
 		TODO("Not yet implemented")