+++ docs/HISTORY.md
@@ -5,6 +5,8 @@ Versions prior to 1.70 are not listed.
 ## 1.85
 
 - Bluetooth device names that appear in UI will now use the user-defined alias name (if available).
+- Added "send initial MIDI messages at title screen" preference. You can now make the app send any
+  any initial MIDI messages *at* the title screen or *after* the title screen.
 
 ## 1.84